Created a script to help locate missing files
Added a new script findfile.ps1 that can accept a list of filenames, or
a directory containing files and then search for them in a directory
structure.

The output produced depends on the options provided but either a full
path to all found files, a unique list of all found filenames or a
unique list of all missing files is output.